Add DIQT replicate nb, start data loader
- What change does this apply?
Adds a directory and jupyter notebook for replicating the
Blumber.etal.2018 DIQT paper. Makes a general layout structure for said
notebook. Starts the data loader for dMRI data from HCP using `torchio`.

- Why is this change necessary?
Replication of DIQT is necessary for preliminary results, and should
provide a good starting point for future iterations. And you need a data
loader to load data, so...

- What are the next planned changes?
Complete the data loader with the `torchio.Dataset` class, move onto
using a `pytorch-lightning` trainer object for training and validation.