Add workaround for shape check, v-sized patches
- What change does this apply?
Adds code to work around the torchio limitation that all Images in a
Subject must have the same spatial shape. Changes the "low-res" image
items in the Subject to be a plain `dict`, with fields mapping to
"data", "affine", etc.

Adds variable-sized patch shapes to the `MultiResSampler` class.

- Why is this change necessary?
Upon reading the Tanno, et. al. 2021 paper, I realized that the low-res
patches can cover a larger spatial area than the corresponding high-res
patches. So, that had to be added into the sampler.

- What are the next planned changes?