Update conda env specs
*Note*: The current environment has a custom version of the
`torchvision` package. This is because the current stable branch has a
weird dependency pin on `SimpleITK < 2` which has been resolved in the
nightly builds. However, the nightly builds require updating to the
nightly pytorch builds (1.9), which are not stable. So, I just edited
the conda package file for a compatible version of torchvision until
this dependency pin is removed in the stable branch.

See <https://github.com/pytorch/vision/issues/3207> for details.