Rename src to lib, first fix of Multires sampler
- What change does this apply?
Changes name of `src/` to `lib/`, just seeing how I like it.

Tries to fix the bug in `MultiresSampler` and `MultiresGridSampler` that
ignores the low-res oversampling factor (in relation to the high-res
spatial extent). But, it doesn't seem to work, so a new pipeline will
need to be worked up.

- What are the next planned changes?
Pipeline needs to be changed, the current approach is too complex to
fix current bugs.
    - Use Pad transforms before downsampling, for simplicity.
    - Create a MaskDilateTransform class, so dilation can be added into
    the actual pipeline in the earlier stages.
    - Allow MultiresGridSampler to accept a 3D binary mask tensor.