Add val step, basic tb tracking
- What change does this apply?
Add validation step in the pytorch lightning model.

Make major adjustments to the logging system. Now, utilize tensorboard,
along with an unstructured log text file for misc. info dumping.
Currently, the code will save out hyperparams and other params, record
loss and RMSE metrics at the train and val steps, and save out a small
sub-region of the validation volume for visualization.

Add SubGridAggregator class and a collate_locations function to the
`lib.viz` module. These are for aggregating patches to create a sub-set
of an entire volume, most useful in visualization.